% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_functions.r
\name{NA_compl}
\alias{NA_compl}
\title{NA Completion (xts Vector)}
\usage{
NA_compl(x, trans_flag = 1, wind = 5)
}
\arguments{
\item{x}{data as a vector}

\item{trans_flag}{code containing the transformation flag, currently only 1 (replace with one-sided median of preceding observations)  and 2 (replace with local median (obs + or - 3 on each side, adjusted if at beginning or end)) supported}

\item{wind}{window for median calculation (one-sided)}
}
\value{
transformed vector
}
\description{
NA  completion
}
\details{
series containing NA values are completed with one-sided median (\code{wind} preceding obs) \cr use of two-sided values (3 preceding and 3 following) wouldn't be possible with real time values \cr (but might be useful for factor detection from past values, or real time with delay - for factor detection...)
}
\note{
Use \code{NA_compl_matrix} for matrix (whole data set).
}
\author{
Sorin T. Pascu
}
\seealso{
\code{NA_compl_matrix}
}

